get types right

+++ emulation/lb_rcv.cc
@@ -202,7 +202,7 @@ int main (int argc, char *argv[])
         uint64_t re_tick = NTOHLL(re_tick_net);
 #endif
 
-        int test_pre_swap = 1, test_post_swap = 1;
+        uint64_t test_pre_swap = 1, test_post_swap = 1;
         NTOHLL(test_post_swap);
         if (test_pre_swap != test_post_swap) {
             fprintf ( stdout, "NOTHLL(x) is a problem, pre swap i = 1, post swap i = %d\n", test_post_swap);
